pyproject.toml
[tool.poetry]
name = "kube-compose"
version = "0.1.0"
description = ""
authors = ["Daniel Clarke <u8sand@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
click = "^8.1.7"
pyyaml = "^6.0.1"

[tool.poetry.scripts]
kube-compose = 'kube_compose.__main__:cli'

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
